pyproject.toml
[tool.poetry]
name = "data_parsers"
version = "0.1.0"
description = ""
authors = ["melandur <account@qunit.ch>"]

[tool.poetry.dependencies]
python = "^3.9.13"
dicom2nifti = "^2.4.2"
pydicom = "^2.3.0"
SimpleITK = "2.0.2"
loguru = "^0.6.0"
nibabel = "^4.0.1"
openpyxl = "^3.0.10"
xlrd = "1.2.0"
pyarrow = "^9.0.0"
fastparquet = "^0.8.1"
matplotlib = "^3.5.3"
seaborn = "^0.11.2"
glueviz = "^1.2.0"
PyQt5 = "^5.15.7"
glue-vispy-viewers = "^1.0.5"
glue-plotly = "^0.3"
specviz = "^0.8"
pylint = "^2.15.0"
jupyter = "^1.0.0"
notebook = "^6.4.12"
hydra-core = "^1.3"
omegaconf = "^2.2"
scikit-learn = "^1.2.0"
xlsxwriter = "^3.0.7"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
skip-string-normalization = true

[tool.isort]
profile = "black"
